--- view/app/default/connexion.php
@@ -1,23 +1,0 @@
-<?php
-
-
-use App\Service\Form;
-
-?>
-
-    <h1> Connexion </h1>
-
-<?php
-
-$form = new Form();
-
-echo $form->label('email','Email');
-echo $form->input('email','email');
-echo $form->error('email');
-echo $form->label('Mot de passe','mdp');
-echo $form->input('mdp','password');
-echo $form->error('mdp');
-
-
-echo $form->submit();
-
